+++ src/azure-cli/azure/cli/command_modules/acr/tests/latest/test_acr_commands.py
@@ -31,6 +31,15 @@ class AcrCommandsTests(ScenarioTest):
                          self.check('tags', {'cat': '', 'foo': 'bar'}),
                          self.check('adminUserEnabled', True),
                          self.check('provisioningState', 'Succeeded')])
+                         
+        # test retention
+        self.cmd('acr config retention update -n {} --status enabled --days 30'.format(registry_name),
+                 checks=[self.check('status', "enabled"),
+                         self.check('days', 30)])
+
+        self.cmd('acr config retention show -n {}'.format(registry_name),
+                 checks=[self.check('status', "enabled"),
+                         self.check('days', 30)])
 
         # test content-trust
         self.cmd('acr config content-trust update -n {} --status enabled'.format(registry_name),